mqtt-consumer/requirements.txt
paho-mqtt
datetime
pendulum